python_backend/requirements.txt
fastapi==0.115.6
pydantic==2.10.5
uvicorn==0.34.0
asyncpg==0.30.0
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
cryptography==44.0.0